pyproject.toml
[tool.poetry]
name = "qor"
version = "0.1.0"
description = "`qor` python framework infront of the C `kore` framework"
authors = ["tabebqena <tabebqena@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
watchdog = "^2.2.1"
python-dotenv = "^1.0.0"
click = "^8.1.3"

[tool.poetry.dev-dependencies]
black = "^23.1.0"
requests = "^2.28.2"
locust = "^2.14.2"
Sphinx = "^6.1.3"
poethepoet = "^0.18.1"
docutils = "^0.19"
pytest = "^7.2.1"
pytest-cov = "^4.0.0"
setuptools = "^67.4.0"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
qor = 'qor.cli:qor'


[tool.poe.tasks]
format = "black ."  # strings are interpreted as commands by default
clean = """
# Multiline commands including comments work too. Unescaped whitespace is ignored.
rm -rf .coverage
       .mypy_cache
       .pytest_cache
       dist
       ./**/__pycache__
"""
greet="echo Hello $USER"  # Environment variables work, even though there's no shell!
build-docs=".venv/bin/sphinx-build docs/source/ docs/build"
test="pytest -v -s --cov=. --cov-report html ."